no longer prevent mods from ping
uses hard coded ids atm, will fix later

+++ main.py
@@ -21,7 +21,8 @@ async def on_ready():
 async def on_message(message):
 
     # Ping warning
-    if ("-ping" in message.content) and message.author.id != client.user.id:
+    # change last 2 conditionals to single if have mod role but cba atm
+    if ("-ping" in message.content) and message.author.id != client.user.id and message.author.id != "227187657715875841" and message.author.id != "108875988967882752":
         warningPing = "**Do not abuse the ping role!** " + message.author.mention
         await client.send_message(message.channel, warningPing)
         await client.delete_message(message)
